@@ -5,8 +5,8 @@ buildscript {
     ext {
         // Sdk and tools
         minSdkVersion = 16
-        targetSdkVersion = 28
-        compileSdkVersion = 28
+        targetSdkVersion = 29
+        compileSdkVersion = 29
         buildToolsVersion = '27.0.3'
 
         // App dependencies
@@ -31,20 +31,21 @@ buildscript {
     repositories {
         google()
         jcenter()
-        maven {
-            url 'https://maven.fabric.io/public'
-        }
     }
 
     dependencies {
         classpath "com.android.tools.build:gradle:$gradleVersion"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
         classpath 'com.google.gms:google-services:4.2.0'
-        classpath 'io.fabric.tools:gradle:1.28.0'
         classpath 'de.felixschulze.gradle:gradle-hockeyapp-plugin:3.6'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
+
+        // Add the Crashlytics Gradle plugin (use v2.0.0+ if you built your app with Android Studio 4.1).
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:2.3.0'
+        // Add the dependency for the Performance Monitoring plugin
+        classpath 'com.google.firebase:perf-plugin:1.3.3'  // Performance Monitoring plugin
     }
 }
 
