@@ -82,7 +82,6 @@ import com.bykea.pk.partner.utils.audio.BykeaAmazonClient;
 import com.bykea.pk.partner.widgets.AutoFitFontTextView;
 import com.bykea.pk.partner.widgets.DashedLine;
 import com.bykea.pk.partner.widgets.FontTextView;
-import com.crashlytics.android.Crashlytics;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.maps.CameraUpdate;
@@ -100,6 +99,7 @@ import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.android.gms.maps.model.PolylineOptions;
+import com.google.firebase.crashlytics.FirebaseCrashlytics;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import com.google.maps.android.PolyUtil;
@@ -130,12 +130,15 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
+import static com.bykea.pk.partner.utils.Constants.ACTION;
 import static com.bykea.pk.partner.utils.Constants.ApiError.BUSINESS_LOGIC_ERROR;
 import static com.bykea.pk.partner.utils.Constants.DIGIT_THOUSAND;
 import static com.bykea.pk.partner.utils.Constants.DIGIT_ZERO;
 import static com.bykea.pk.partner.utils.Constants.DIRECTION_API_MIX_THRESHOLD_METERS;
 import static com.bykea.pk.partner.utils.Constants.MAX_LIMIT_LOAD_BOARD;
+import static com.bykea.pk.partner.utils.Constants.MSG;
 import static com.bykea.pk.partner.utils.Constants.PLUS;
+import static com.bykea.pk.partner.utils.Constants.ServiceCode.DISPATCH_RIDE;
 import static com.bykea.pk.partner.utils.Constants.ServiceCode.MART;
 import static com.bykea.pk.partner.utils.Constants.ServiceCode.OFFLINE_DELIVERY;
 import static com.bykea.pk.partner.utils.Constants.ServiceCode.OFFLINE_RIDE;
@@ -882,7 +885,8 @@ public class BookingActivity extends BaseActivity implements GoogleApiClient.OnC
                         ActivityStackManager.getInstance()
                                 .startChatActivity(callData.getPassName(), "", true, mCurrentActivity);
                     } else {
-                        if (callData.getCreator_type().toUpperCase().equalsIgnoreCase(Constants.IOS))
+                        if (callData.getCreator_type().toUpperCase().equalsIgnoreCase(Constants.IOS) ||
+                                (callData.getServiceCode() != null && callData.getServiceCode() == DISPATCH_RIDE))
                             Utils.sendSms(mCurrentActivity, callData.getPhoneNo());
                         else
                             Utils.sendSms(mCurrentActivity, callData.getSenderPhone());
@@ -1202,25 +1206,25 @@ public class BookingActivity extends BaseActivity implements GoogleApiClient.OnC
             mCurrentActivity.runOnUiThread(new Runnable() {
                 @Override
                 public void run() {
-                    if (intent.getStringExtra("action").equalsIgnoreCase(Keys.BROADCAST_CANCEL_RIDE) ||
-                            intent.getStringExtra("action").equalsIgnoreCase(Keys.BROADCAST_CANCEL_BATCH)) {
+                    if (intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.BROADCAST_CANCEL_RIDE) ||
+                            intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.BROADCAST_CANCEL_BATCH)) {
                         cancelByPassenger(false);
                     }
-                    if (intent.getStringExtra("action").equalsIgnoreCase(Keys.BROADCAST_CANCEL_BY_ADMIN)) {
-                        String message = intent.getStringExtra("msg");
+                    if (intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.BROADCAST_CANCEL_BY_ADMIN)) {
+                        String message = intent.getStringExtra(MSG);
                         cancelByPassenger(true);
                     }
-                    if (intent.getStringExtra("action").equalsIgnoreCase(Keys.BROADCAST_COMPLETE_BY_ADMIN)) {
+                    if (intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.BROADCAST_COMPLETE_BY_ADMIN)) {
                         playNotificationSound();
-                        onCompleteByAdmin(intent.getStringExtra("msg"));
+                        onCompleteByAdmin(intent.getStringExtra(MSG));
                     }
-                    if (intent.getStringExtra("action").equalsIgnoreCase(Keys.BROADCAST_DROP_OFF_UPDATED)) {
+                    if (intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.BROADCAST_DROP_OFF_UPDATED)) {
                         playNotificationSound();
                         Utils.appToast(getString(R.string.drop_off_update_by_passenger));
                         shouldRefreshDropOffMarker = true;
                         dataRepository.requestRunningTrip(mCurrentActivity, handler);
                     }
-                    if (intent.getStringExtra("action").equalsIgnoreCase(Keys.TRIP_DATA_UPDATED)) {
+                    if (intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.TRIP_DATA_UPDATED)) {
                         playNotificationSound();
                         Utils.appToast("Trip Details has been Added by Passenger.");
                         callData = AppPreferences.getCallData();
@@ -1231,7 +1235,7 @@ public class BookingActivity extends BaseActivity implements GoogleApiClient.OnC
                             showDropOffPersonInfo();
                         showWalletAmount();
                     }
-                    if (intent.getStringExtra("action").equalsIgnoreCase(Keys.BROADCAST_BATCH_UPDATED)) {
+                    if (intent.getStringExtra(ACTION).equalsIgnoreCase(Keys.BROADCAST_BATCH_UPDATED)) {
                         dataRepository.requestRunningTrip(mCurrentActivity, handler);
                     }
                 }
@@ -1988,8 +1992,11 @@ public class BookingActivity extends BaseActivity implements GoogleApiClient.OnC
 
         mGoogleMap.clear();
         driverMarker = null;
-        updateDriverMarker(mCurrentLocation.getLatitude() + "",
-                mCurrentLocation.getLongitude() + "");
+
+        if (mCurrentLocation != null) {
+            updateDriverMarker(mCurrentLocation.getLatitude() + "",
+                    mCurrentLocation.getLongitude() + "");
+        }
 
         if (callData.getPickupStop() != null && StringUtils.isNotEmpty(callData.getStartLat()) && StringUtils.isNotEmpty(callData.getStartLng())) {
             if (callData.getStatus().equalsIgnoreCase(TripStatus.ON_ACCEPT_CALL) /*||
@@ -2729,9 +2736,9 @@ public class BookingActivity extends BaseActivity implements GoogleApiClient.OnC
             @Override
             public void onJobFinished(@NotNull FinishJobResponseData data, @NotNull String request, @NotNull String resp) {
                 AppPreferences.removeReceivedMessageCount();
-                Crashlytics.setUserIdentifier(AppPreferences.getPilotData().getId());
-                Crashlytics.setString("Finish Job Request Trip ID", callData.getTripId());
-                Crashlytics.setString("Finish Job Response", resp);
+                FirebaseCrashlytics.getInstance().setUserId(AppPreferences.getPilotData().getId());
+                FirebaseCrashlytics.getInstance().setCustomKey("Finish Job Request Trip ID", callData.getTripId());
+                FirebaseCrashlytics.getInstance().setCustomKey("Finish Job Response", resp);
 
                 onFinished(data, endAddress);
             }
