@@ -199,7 +199,7 @@ public class Constants {
         public static final String NOTIFICATION_CHANNEL_ID = "bykea_p_channel_id_for_loc";
         public static final String NOTIFICATION_CHANNEL_NAME = "Bykea Active/Inactive Status";
         public static final String NOTIFICATION_CONTENT_TITLE = "Bykea Partner";
-
+        public static final String CONTENT_AVAILABLE_CHANNEL = "content-available-Channel";
         public static final String DATA_TYPE_TELLO_VAL = "content-available-IM";
         public static final String DATA_TYPE_TELLO = "jobCode";
         public static final String EVENT_TYPE = "event";
